requirements.txt
django>=3.1
django-cors-middleware>=1.3.1
django-debug-toolbar==2.0
django-currentuser>=0.5.0
django-cors-headers>=3.2.1
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.3.0
django-crispy-forms>=1.9.0
sentry-sdk==0.14.4
celery>=4.4.5
django-celery-results>=1.2.1
django-redis>=4.12.1
redis>=3.5.3
future>=0.18.2
eventlet>=0.25.2
pyotp>=2.3.0
gunicorn==19.9.0
psycopg2>=2.7.7
psycopg2-binary>=2.7.7
Pillow>=6.2.1
whitenoise>=3.2
honcho>=1.0.1
python-memcached>=1.59